# openapi.yaml
openapi: 3.0.3
info:
  title: HLS Streaming Server API
  description: |
    Complete REST API for managing HLS streaming channels, media libraries, and dynamic playlists.

    ## Features
    - Multi-channel HLS streaming with FFmpeg
    - Dynamic playlist generation via schedule blocks
    - Jellyfin-style library management
    - Media bucket organization
    - EPG/XMLTV generation
    - IPTV M3U playlist export
    - Session-based authentication
    - Virtual time tracking for 24/7 simulation

    ## Authentication
    The API supports two authentication methods:
    1. **API Key** (Legacy): Pass via `X-API-Key` header or `apiKey` query parameter
    2. **Session Token** (Modern): Pass via `Authorization: Bearer <token>` header or `sessionToken` cookie

    Most endpoints require authentication except for streaming and public EPG endpoints.

  version: 1.0.0
  contact:
    name: HLS Streaming Server
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://your-server.com
    description: Production server
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
    CookieAuth:
      type: apiKey
      in: cookie
      name: sessionToken
...
paths:
...
  /{slug}/master.m3u8:
    get:
      tags: [Streaming]
      summary: HLS master playlist
      description: Returns the master HLS playlist for a channel
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Master playlist
          content:
            application/vnd.apple.mpegurl:
              schema:
                type: string
                example: |
                  #EXTM3U
                  #EXT-X-STREAM-INF:BANDWIDTH=1628000,RESOLUTION=1920x1080
                  stream.m3u8
  /{slug}/stream.m3u8:
    get:
      tags: [Streaming]
      summary: HLS media playlist
      description: Returns the media playlist with segment references
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Media playlist
          content:
            application/vnd.apple.mpegurl:
              schema:
                type: string
...